# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/rents": {
            "get": {
                "description": "list rent entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List rent entities",
                "operationId": "list-rent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Rent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create rent",
                "operationId": "create-rent",
                "parameters": [
                    {
                        "description": "Rent entity",
                        "name": "rent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Rent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Rent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/rents/{id}": {
            "get": {
                "description": "get rent by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a rent entity by ID",
                "operationId": "get-rent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Rent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update rent by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a rent entity by ID",
                "operationId": "update-rent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rent entity",
                        "name": "rent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Rent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Rent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get rent by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a rent entity by ID",
                "operationId": "delete-rent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Rent": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "string"
                },
                "kinName": {
                    "type": "string"
                },
                "kinTel": {
                    "type": "string"
                },
                "nurse": {
                    "type": "integer"
                },
                "patient": {
                    "type": "integer"
                },
                "rentID": {
                    "type": "string"
                },
                "room": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Bloodtype": {
            "type": "object",
            "properties": {
                "BTname": {
                    "description": "BTname holds the value of the \"BTname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BloodtypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BloodtypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BloodtypeEdges": {
            "type": "object",
            "properties": {
                "frombloodtype": {
                    "description": "Frombloodtype holds the value of the frombloodtype edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Patient"
                    }
                }
            }
        },
...
        "ent.CoveredPerson": {
            "type": "object",
            "properties": {
                "certificate_id": {
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CoveredPersonQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CoveredPersonEdges"
                },
                "fund_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "patient_id": {
                    "type": "integer"
                },
                "schemeType_id": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Diagnose": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DiagnoseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DiagnoseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Gender": {
            "type": "object",
            "properties": {
                "Gname": {
                    "description": "Gname holds the value of the \"Gname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "nurse_email": {
                    "description": "NurseEmail holds the value of the \"nurse_email\" field.",
                    "type": "string"
                },
                "nurse_name": {
                    "description": "NurseName holds the value of the \"nurse_name\" field.",
                    "type": "string"
                },
                "nurse_password": {
                    "description": "NursePassword holds the value of the \"nurse_password\" field.",
                    "type": "string"
                },
                "nurse_tel": {
                    "description": "NurseTel holds the value of the \"nurse_tel\" field.",
                    "type": "string"
                }
            }
        },
        "ent.NurseEdges": {
            "type": "object",
            "properties": {
                "fromnurse": {
                    "description": "Fromnurse holds the value of the fromnurse edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Rent"
                    }
                },
                "nurseOperativerecord": {
                    "description": "NurseOperativerecord holds the value of the Nurse_Operativerecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                },
                "nursePrescription": {
                    "description": "NursePrescription holds the value of the nurse_prescription edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                }
            }
        },
...
        "ent.Operativerecord": {
            "type": "object",
            "properties": {
                "OperativeTime": {
                    "description": "OperativeTime holds the value of the \"OperativeTime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OperativerecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OperativerecordEdges"
                },
                "examinationroom_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "nurse_id": {
                    "type": "integer"
                },
                "operative_id": {
                    "type": "integer"
                },
                "tool_id": {
                    "type": "integer"
                }
            }
        },
        "ent.OperativerecordEdges": {
            "type": "object",
            "properties": {
                "examinationroom": {
                    "description": "Examinationroom holds the value of the Examinationroom edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Examinationroom"
                },
                "nurse": {
                    "description": "Nurse holds the value of the Nurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "operative": {
                    "description": "Operative holds the value of the Operative edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Operative"
                },
                "tool": {
                    "description": "Tool holds the value of the Tool edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Tool"
                }
            }
        },
        "ent.Patient": {
            "type": "object",
            "properties": {
                "Patient_age": {
                    "description": "PatientAge holds the value of the \"Patient_age\" field.",
                    "type": "integer"
                },
                "Patient_height": {
                    "description": "PatientHeight holds the value of the \"Patient_height\" field.",
                    "type": "number"
                },
                "Patient_name": {
                    "description": "PatientName holds the value of the \"Patient_name\" field.",
                    "type": "string"
                },
                "Patient_weight": {
                    "description": "PatientWeight holds the value of the \"Patient_weight\" field.",
                    "type": "number"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PatientEdges": {
            "type": "object",
            "properties": {
                "bloodtype": {
                    "description": "Bloodtype holds the value of the bloodtype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bloodtype"
                },
                "frompatient": {
                    "description": "Frompatient holds the value of the frompatient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Rent"
                },
                "gender": {
                    "description": "Gender holds the value of the gender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "patientCoveredPerson": {
                    "description": "PatientCoveredPerson holds the value of the Patient_CoveredPerson edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.CoveredPerson"
                    }
                },
                "patientDiagnose": {
                    "description": "PatientDiagnose holds the value of the patient_diagnose edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Diagnose"
                    }
                },
                "patientPrescription": {
                    "description": "PatientPrescription holds the value of the patient_prescription edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                },
                "prefix": {
                    "description": "Prefix holds the value of the prefix edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prefix"
                }
            }
        },
        "ent.Prefix": {
            "type": "object",
            "properties": {
                "Pname": {
                    "description": "Pname holds the value of the \"Pname\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrefixQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrefixEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Prescription": {
            "type": "object",
            "properties": {
                "Prescrip_DateTime": {
                    "description": "PrescripDateTime holds the value of the \"Prescrip_DateTime\" field.",
                    "type": "string"
                },
                "Prescrip_Note": {
                    "description": "PrescripNote holds the value of the \"Prescrip_Note\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrescriptionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrescriptionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PrescriptionEdges": {
            "type": "object",
            "properties": {
                "doctor": {
                    "description": "Doctor holds the value of the doctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "drug": {
                    "description": "Drug holds the value of the drug edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Drug"
                },
                "nurse": {
                    "description": "Nurse holds the value of the nurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                }
            }
        },
        "ent.Rent": {
            "type": "object",
            "properties": {
                "added_time": {
                    "description": "AddedTime holds the value of the \"added_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "kin_name": {
                    "description": "KinName holds the value of the \"kin_name\" field.",
                    "type": "string"
                },
                "kin_tel": {
                    "description": "KinTel holds the value of the \"kin_tel\" field.",
                    "type": "string"
                },
                "patient_id": {
                    "type": "integer"
                },
                "rent_id": {
                    "description": "RentID holds the value of the \"rent_id\" field.",
                    "type": "string"
                }
            }
        },
        "ent.RentEdges": {
            "type": "object",
            "properties": {
                "nurse": {
                    "description": "Nurse holds the value of the nurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Room"
                }
            }
        },
        "ent.Room": {
            "type": "object",
            "properties": {
                "building": {
                    "description": "Building holds the value of the \"building\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomEdges"
                },
                "floor": {
                    "description": "Floor holds the value of the \"floor\" field.",
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.RoomEdges": {
            "type": "object",
            "properties": {
                "rents": {
                    "description": "Rents holds the value of the rents edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Rent"
                },
                "roomtype": {
                    "description": "Roomtype holds the value of the roomtype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Roomtype"
                }
            }
        },
        "ent.Roomtype": {
            "type": "object",
            "properties": {
                "areasize": {
                    "description": "Areasize holds the value of the \"areasize\" field.",
                    "type": "number"
                },
                "bathroom": {
                    "description": "Bathroom holds the value of the \"bathroom\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomtypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomtypeEdges"
                },
                "etc": {
                    "description": "Etc holds the value of the \"etc\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "toilets": {
                    "description": "Toilets holds the value of the \"toilets\" field.",
                    "type": "integer"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}